# src/lib/openapi.yml
openapi: 3.0.0
info:
  title: Turnip React API
  version: 0.6.0
servers:
  - url: https://react.turnipxenon.com
    description: Production server
  - url: http://localhost:3000
    description: Local test server
components:
  schemas:
...
    JobEvent:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        anonymizedNotes:
          type: string
        publicNotes:
          type: string
        eventDate:
          type: string
        jobPostingId:
          type: string
    JobPosting:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        alias:
          type: string
        company:
          type: string
        jobTitle:
          type: string
        jobLink:
          type: string
        resumeLink:
          type: string
        startDate:
          type: string
        sessionId:
          type: string
        userId:
          type: string
        notes:
          type: string
        status:
          type: string
          enum:
            - Queued
            - Archived
            - Waiting
            - Active
            - Rejected
        events:
          type: array
          items:
            $ref: "#/components/schemas/JobEvent"
...
  parameters:
...
    PostingParameterID:
      name: postingId   # Note the name is the same as in the path
      in: path
      required: true
      description: The job posting ID
      schema:
        type: string
...
paths:
...
  /api/job-tracker/postings/{postingId}:
    get:
      summary: Get job posting
      parameters:
        - $ref: "#/components/parameters/PostingParameterID"
      responses:
        '200':
          description: Job posting found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
        '401':
          description: Unauthorized error
        '404':
          description: Missing job posting
    put:
      summary: Update job posting
      parameters:
        - $ref: "#/components/parameters/PostingParameterID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPosting"
      responses:
        '200':
          description: Job posting updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPosting"
        '401':
          description: Unauthorized error
        '404':
          description: Missing job posting
    delete:
      summary: Delete job posting
      parameters:
        - $ref: "#/components/parameters/PostingParameterID"
      responses:
        '200':
          description: Successfully delete job session
        '401':
          description: Unauthorized error
        '404':
          description: Missing job posting
...